Create endpoints to deal with Slack API
When users within dinner rotation interact with the messages
sent out to the channel (i.e. liking the 'who wants to eat today post'),
the backend is responsible for 'logging' that information and reflecting
that information by making changes to the database. Additionally,
the backend creates interactive messages to make it clear to the user
that changes happened in the background (i.e. saying 'acknowledged' when
the user clicks 'yes' to join dinner rotation that semester)